+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Add ui chunk profiling data category. ([#4593](https://github.com/getsentry/relay/pull/4593))
 - Switch global rate limiter to a service. ([#4581](https://github.com/getsentry/relay/pull/4581))
+- Extract transaction breakdowns into measurements. ([#4599](https://github.com/getsentry/relay/pull/4599))
 
 ## 25.3.0